requirements.txt
streamlit
replicate
Pillow